# Remove array allocation from multi-span header parsing