--- a/src/Servers/Kestrel/perf/Microbenchmarks/HttpParserBenchmark.cs
+++ b/src/Servers/Kestrel/perf/Microbenchmarks/HttpParserBenchmark.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Buffers;
+using System.IO.Pipelines;
 using BenchmarkDotNet.Attributes;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
 
@@ -12,36 +13,51 @@ public class HttpParserBenchmark : IHttpRequestLineHandler, IHttpHeadersHandler
     private readonly HttpParser<Adapter> _parser = new HttpParser<Adapter>();
 
     private ReadOnlySequence<byte> _buffer;
+    private ReadOnlySequence<byte> _multispanHeader;
 
-    [Benchmark(Baseline = true, OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
-    public void PlaintextTechEmpower()
+    [GlobalSetup]
+    public void Setup()
     {
-        for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
-        {
-            InsertData(RequestParsingData.PlaintextTechEmpowerRequest);
-            ParseData();
-        }
+        var segment = new BufferSegment();
+        var split = RequestParsingData.UnicodeRequest.Length / 2;
+        segment.SetOwnedMemory(RequestParsingData.UnicodeRequest.AsSpan(0, split).ToArray());
+        segment.End = split;
+        var next = new BufferSegment();
+        next.SetOwnedMemory(RequestParsingData.UnicodeRequest.AsSpan(split).ToArray());
+        next.End = split;
+        segment.SetNext(next);
+        _multispanHeader = new ReadOnlySequence<byte>(segment, 0, next, next.Memory.Length);
     }
 
-    [Benchmark(OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
-    public void JsonTechEmpower()
-    {
-        for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
-        {
-            InsertData(RequestParsingData.JsonTechEmpowerRequest);
-            ParseData();
-        }
-    }
+    //[Benchmark(Baseline = true, OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
+    //public void PlaintextTechEmpower()
+    //{
+    //    for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
+    //    {
+    //        InsertData(RequestParsingData.PlaintextTechEmpowerRequest);
+    //        ParseData();
+    //    }
+    //}
 
-    [Benchmark(OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
-    public void LiveAspNet()
-    {
-        for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
-        {
-            InsertData(RequestParsingData.LiveaspnetRequest);
-            ParseData();
-        }
-    }
+    //[Benchmark(OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
+    //public void JsonTechEmpower()
+    //{
+    //    for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
+    //    {
+    //        InsertData(RequestParsingData.JsonTechEmpowerRequest);
+    //        ParseData();
+    //    }
+    //}
+
+    //[Benchmark(OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
+    //public void LiveAspNet()
+    //{
+    //    for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
+    //    {
+    //        InsertData(RequestParsingData.LiveaspnetRequest);
+    //        ParseData();
+    //    }
+    //}
 
     [Benchmark(OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
     public void Unicode()
@@ -49,6 +65,16 @@ public class HttpParserBenchmark : IHttpRequestLineHandler, IHttpHeadersHandler
         for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
         {
             InsertData(RequestParsingData.UnicodeRequest);
+            ParseData();
+        }
+    }
+
+    [Benchmark(OperationsPerInvoke = RequestParsingData.InnerLoopCount)]
+    public void MultispanUnicodeHeader()
+    {
+        for (var i = 0; i < RequestParsingData.InnerLoopCount; i++)
+        {
+            _buffer = _multispanHeader;
             ParseData();
         }
     }