Noticed this allocation while looking at a memory profile for CertAuth. My hypothesis is that the certificate/tls handshake is pushing the header bytes across a PipeSegment boundary which goes through the multi-span header parsing path, and that path would allocate a `byte[]` if the header crossed multiple spans. This change avoids the allocation by using `stackalloc` or `ArrayPool<byte>.Shared`.

![remove_alloc](https://user-images.githubusercontent.com/7574801/201448592-86bdcefc-c00b-4d0d-bbea-4eab8940cc4f.png)
The highlighted line is now gone from the benchmark 😃 

**Before:**
|                 Method |     Mean |    Error |   StdDev |   Median |        Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|---------:|---------:|---------:|------------:|------:|------:|------:|----------:|
|                Unicode | 371.2 ns | 11.94 ns | 35.22 ns | 356.3 ns | 2,693,742.3 |     - |     - |     - |         - |
| MultispanUnicodeHeader | 573.8 ns | 18.61 ns | 54.87 ns | 552.9 ns | 1,742,893.2 |     - |     - |     - |      48 B |

**After:**
|                 Method |     Mean |    Error |   StdDev |   Median |        Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|---------:|---------:|---------:|------------:|------:|------:|------:|----------:|
|                Unicode | 351.9 ns | 10.64 ns | 31.03 ns | 339.3 ns | 2,841,974.0 |     - |     - |     - |         - |
| MultispanUnicodeHeader - First Commit | 553.0 ns | 16.74 ns | 48.82 ns | 533.3 ns | 1,808,160.4 |     - |     - |     - |         - |
| MultispanUnicodeHeader - Second Commit | 484.9 ns | 11.75 ns | 33.90 ns | 471.2 ns | 2,062,450.8 |     - |     - |     - |         - |

The Op/s difference is noise, especially since `Unicode` shouldn't hit the changed code path.