@BrennanConroy 2022-11-15
I was able to improve `MultispanUnicodeHeader` to 2m Op/s locally by inlining the `PositionOfAny` method, skipping the first span (since it was checked by the caller), and doing operations with the `ReadOnlyMemory` returned from `ReadOnlySequence.TryGet` instead of all the `GetPosition` APIs on `ReadOnlySequence`. The code is a bit yuckier though, so not sure it's worth it.